+++ lib/freeclimb/models/platform_error.rb
@@ -12,46 +12,38 @@ require "date"
 require "time"
 
 module Freeclimb
-  # The `Say` command provides Text-To-Speech (TTS) support. It converts text to speech and then renders it in a female voice back to the caller. `Say` is useful in cases where it's difficult to pre-record a prompt for any reason. `Say` does not allow barge-in unless nested within a `GetSpeech` command. The file will always be played to completion unless nested.
-  class Say < PerclCommand
-    # The message to be played to the caller using TTS. The size of the string is limited to 4 KB (or 4,096 bytes). An empty string will cause the command to be skipped.
-    attr_accessor :text
+  # Standard error structure returned by platform.
+  class PlatformError
+    attr_accessor :code
 
-    # Language and (by implication) the locale to use. This implies the accent and pronunciations to be usde for the TTS. The complete list of valid values for the language attribute is shown below.
-    attr_accessor :language
+    attr_accessor :call
 
-    attr_accessor :engine
+    attr_accessor :url
 
-    # Number of times the text is said. Specifying '0' causes the `Say` action to loop until the Call is hung up.
-    attr_accessor :loop
-
-    # Parameter `privacyMode` will not log the `text` as required by PCI compliance.
-    attr_accessor :privacy_mode
+    attr_accessor :details
 
     # Attribute mapping from ruby-style variable name to JSON key.
     def self.attribute_map
       {
-        text: :text,
-        language: :language,
-        engine: :engine,
-        loop: :loop,
-        privacy_mode: :privacyMode
+        code: :code,
+        call: :call,
+        url: :url,
+        details: :details
       }
     end
 
-    # Returns all the JSON keys this model knows about, including the ones defined in its parent(s)
+    # Returns all the JSON keys this model knows about
     def self.acceptable_attributes
-      attribute_map.values.concat(superclass.acceptable_attributes)
+      attribute_map.values
     end
 
     # Attribute type mapping.
     def self.openapi_types
       {
-        text: :String,
-        language: :String,
-        engine: :TTSEngine,
-        loop: :Integer,
-        privacy_mode: :Boolean
+        code: :Integer,
+        call: :String,
+        url: :String,
+        details: :Object
       }
     end
 
@@ -60,70 +52,48 @@ module Freeclimb
       Set.new([])
     end
 
-    # List of class defined in allOf (OpenAPI v3)
-    def self.openapi_all_of
-      [
-        :PerclCommand
-      ]
-    end
-
     # Initializes the object
     # @param [Hash] attributes Model attributes in the form of hash
     def initialize(attributes = {})
       if !attributes.is_a?(Hash)
-        fail ArgumentError, "The input argument (attributes) must be a hash in `Freeclimb::Say` initialize method"
+        fail ArgumentError, "The input argument (attributes) must be a hash in `Freeclimb::PlatformError` initialize method"
       end
 
       # check to see if the attribute exists and convert string to symbol for hash key
       attributes = attributes.each_with_object({}) { |(k, v), h|
         if !self.class.attribute_map.key?(k.to_sym)
-          fail ArgumentError, "`#{k}` is not a valid attribute in `Freeclimb::Say`. Please check the name to make sure it's valid. List of attributes: " + self.class.attribute_map.keys.inspect
+          fail ArgumentError, "`#{k}` is not a valid attribute in `Freeclimb::PlatformError`. Please check the name to make sure it's valid. List of attributes: " + self.class.attribute_map.keys.inspect
         end
         h[k.to_sym] = v
       }
 
-      # call parent's initialize
-      super
-
-      if attributes.key?(:text)
-        self.text = attributes[:text]
+      if attributes.key?(:code)
+        self.code = attributes[:code]
       end
 
-      if attributes.key?(:language)
-        self.language = attributes[:language]
+      if attributes.key?(:call)
+        self.call = attributes[:call]
       end
 
-      if attributes.key?(:engine)
-        self.engine = attributes[:engine]
+      if attributes.key?(:url)
+        self.url = attributes[:url]
       end
 
-      self.loop = if attributes.key?(:loop)
-        attributes[:loop]
-      else
-        1
+      if attributes.key?(:details)
+        self.details = attributes[:details]
       end
-
-      if attributes.key?(:privacy_mode)
-        self.privacy_mode = attributes[:privacy_mode]
-      end
-      self.command = "Say"
     end
 
     # Show invalid properties with the reasons. Usually used together with valid?
     # @return Array for valid properties with the reasons
     def list_invalid_properties
-      invalid_properties = super
-      if @text.nil?
-        invalid_properties.push('invalid value for "text", text cannot be nil.')
-      end
-
-      invalid_properties
+      []
     end
 
     # Check to see if the all the properties in the model are valid
     # @return true if the model is valid
     def valid?
-      if @text.nil?
+      if @code.nil?
         false
       else
         list_invalid_properties.length == 0
@@ -135,11 +105,10 @@ module Freeclimb
     def ==(other)
       return true if equal?(other)
       self.class == other.class &&
-        text == other.text &&
-        language == other.language &&
-        engine == other.engine &&
-        loop == other.loop &&
-        privacy_mode == other.privacy_mode && super
+        code == other.code &&
+        call == other.call &&
+        url == other.url &&
+        details == other.details
     end
 
     # @see the `==` method
@@ -151,7 +120,7 @@ module Freeclimb
     # Calculates hash code according to all attributes.
     # @return [Integer] Hash code
     def hash
-      [text, language, engine, loop, privacy_mode].hash
+      [code, call, url, details].hash
     end
 
     # Builds the object from hash
@@ -166,7 +135,6 @@ module Freeclimb
     # @return [Object] Returns the model itself
     def build_from_hash(attributes)
       return nil unless attributes.is_a?(Hash)
-      super
       self.class.openapi_types.each_pair do |key, type|
         if attributes[self.class.attribute_map[key]].nil? && self.class.openapi_nullable.include?(key)
           send(:"#{key}=", nil)
@@ -242,7 +210,7 @@ module Freeclimb
     # Returns the object in the form of hash
     # @return [Hash] Returns the object in the form of hash
     def to_hash
-      hash = super
+      hash = {}
       self.class.attribute_map.each_pair do |attr, param|
         value = send(attr)
         if value.nil?