+++ spec/models/incoming_number_result_spec.rb
@@ -116,20 +116,6 @@ describe Freeclimb::IncomingNumberResult do
     end
   end
 
-  describe 'test attribute "voice_enabled"' do
-    it "should work" do
-      instance.voice_enabled = false
-      expect(instance.voice_enabled).to eq(false)
-    end
-  end
-
-  describe 'test attribute "sms_enabled"' do
-    it "should work" do
-      instance.sms_enabled = false
-      expect(instance.sms_enabled).to eq(false)
-    end
-  end
-
   describe 'test attribute "offnet"' do
     it "should work" do
       instance.offnet = false
@@ -174,10 +160,6 @@ describe Freeclimb::IncomingNumberResult do
 
           country: "TS",
 
-          voice_enabled: true,
-
-          sms_enabled: true,
-
           offnet: true,
 
           tfn: TFN.new
@@ -212,10 +194,6 @@ describe Freeclimb::IncomingNumberResult do
           region: "TS",
 
           country: "TS",
-
-          voice_enabled: true,
-
-          sms_enabled: true,
 
           offnet: true,
 
@@ -253,10 +231,6 @@ describe Freeclimb::IncomingNumberResult do
           region: "TS",
 
           country: "TS",
-
-          voice_enabled: true,
-
-          sms_enabled: true,
 
           offnet: true,
 
@@ -297,10 +271,6 @@ describe Freeclimb::IncomingNumberResult do
 
         country: "TS",
 
-        voice_enabled: true,
-
-        sms_enabled: true,
-
         offnet: true,
 
         tfn: TFN.new
@@ -317,9 +287,9 @@ describe Freeclimb::IncomingNumberResult do
 
   describe 'test method "eql?"' do
     it "checks if objects are equal" do
-      obj_Capabilities = Capabilities.new
+      obj_capabilities = Object.new
 
-      obj_TFN = TFN.new
+      obj_tfn = Object.new
 
       instance_1 = Freeclimb::IncomingNumberResult.new(
         uri: "TS",
@@ -330,7 +300,7 @@ describe Freeclimb::IncomingNumberResult do
 
         revision: 1,
 
-        capabilities: obj_Capabilities,
+        capabilities: obj_capabilities,
 
         campaign_id: "TS",
 
@@ -348,13 +318,9 @@ describe Freeclimb::IncomingNumberResult do
 
         country: "TS",
 
-        voice_enabled: true,
-
-        sms_enabled: true,
-
         offnet: true,
 
-        tfn: obj_TFN
+        tfn: obj_tfn
       )
       instance_2 = Freeclimb::IncomingNumberResult.new(
         uri: "TS",
@@ -365,7 +331,7 @@ describe Freeclimb::IncomingNumberResult do
 
         revision: 1,
 
-        capabilities: obj_Capabilities,
+        capabilities: obj_capabilities,
 
         campaign_id: "TS",
 
@@ -383,13 +349,9 @@ describe Freeclimb::IncomingNumberResult do
 
         country: "TS",
 
-        voice_enabled: true,
-
-        sms_enabled: true,
-
         offnet: true,
 
-        tfn: obj_TFN
+        tfn: obj_tfn
       )
       expect(instance_1.eql?(instance_2)).to eq(true)
     end
@@ -422,10 +384,6 @@ describe Freeclimb::IncomingNumberResult do
 
         country: "TS",
 
-        voice_enabled: true,
-
-        sms_enabled: true,
-
         offnet: true,
 
         tfn: TFN.new
@@ -456,10 +414,6 @@ describe Freeclimb::IncomingNumberResult do
         region: "ST",
 
         country: "ST",
-
-        voice_enabled: false,
-
-        sms_enabled: false,
 
         offnet: false,
 
@@ -498,10 +452,6 @@ describe Freeclimb::IncomingNumberResult do
 
         country: "TS",
 
-        voice_enabled: true,
-
-        sms_enabled: true,
-
         offnet: true,
 
         tfn: TFN.new
@@ -539,10 +489,6 @@ describe Freeclimb::IncomingNumberResult do
 
         country: "TS",
 
-        voice_enabled: true,
-
-        sms_enabled: true,
-
         offnet: true,
 
         tfn: TFN.new
@@ -573,10 +519,6 @@ describe Freeclimb::IncomingNumberResult do
         region: "TS",
 
         country: "TS",
-
-        voice_enabled: true,
-
-        sms_enabled: true,
 
         offnet: true,
 
@@ -614,10 +556,6 @@ describe Freeclimb::IncomingNumberResult do
       region: "TS",
 
       country: "TS",
-
-      voice_enabled: true,
-
-      sms_enabled: true,
 
       offnet: true,
 
@@ -675,14 +613,6 @@ describe Freeclimb::IncomingNumberResult do
       expect(instance._deserialize("String", instance.country)).to be_a_kind_of(String)
     end
 
-    it "deserializes the data of voice_enabled" do
-      expect(instance._deserialize("Boolean", instance.voice_enabled)).to be_a_kind_of(TrueClass)
-    end
-
-    it "deserializes the data of sms_enabled" do
-      expect(instance._deserialize("Boolean", instance.sms_enabled)).to be_a_kind_of(TrueClass)
-    end
-
     it "deserializes the data of offnet" do
       expect(instance._deserialize("Boolean", instance.offnet)).to be_a_kind_of(TrueClass)
     end
@@ -721,10 +651,6 @@ describe Freeclimb::IncomingNumberResult do
 
         country: "TS",
 
-        voice_enabled: true,
-
-        sms_enabled: true,
-
         offnet: true,
 
         tfn: TFN.new
@@ -761,10 +687,6 @@ describe Freeclimb::IncomingNumberResult do
         region: "TS",
 
         country: "TS",
-
-        voice_enabled: true,
-
-        sms_enabled: true,
 
         offnet: true,
 
@@ -804,10 +726,6 @@ describe Freeclimb::IncomingNumberResult do
 
         country: "TS",
 
-        voice_enabled: true,
-
-        sms_enabled: true,
-
         offnet: true,
 
         tfn: obj
@@ -838,10 +756,6 @@ describe Freeclimb::IncomingNumberResult do
         region: "TS",
 
         country: "TS",
-
-        voice_enabled: true,
-
-        sms_enabled: true,
 
         offnet: true,
 
@@ -876,10 +790,6 @@ describe Freeclimb::IncomingNumberResult do
       region: "TS",
 
       country: "TS",
-
-      voice_enabled: true,
-
-      sms_enabled: true,
 
       offnet: true
     )
@@ -921,12 +831,6 @@ describe Freeclimb::IncomingNumberResult do
     end
     it "returns country in the form of hash" do
       expect(instance._to_hash(instance.country)).to eq(instance.country)
-    end
-    it "returns voice_enabled in the form of hash" do
-      expect(instance._to_hash(instance.voice_enabled)).to eq(instance.voice_enabled)
-    end
-    it "returns sms_enabled in the form of hash" do
-      expect(instance._to_hash(instance.sms_enabled)).to eq(instance.sms_enabled)
     end
     it "returns offnet in the form of hash" do
       expect(instance._to_hash(instance.offnet)).to eq(instance.offnet)