+++ spec/models/queue_result_spec.rb
@@ -231,7 +231,7 @@ describe Freeclimb::QueueResult do
 
   describe 'test method "eql?"' do
     it "checks if objects are equal" do
-      obj_Object = Object.new
+      obj_subresource_uris = Object.new
 
       instance_1 = Freeclimb::QueueResult.new(
         uri: "TS",
@@ -254,7 +254,7 @@ describe Freeclimb::QueueResult do
 
         average_queue_removal_time: 1,
 
-        subresource_uris: obj_Object
+        subresource_uris: obj_subresource_uris
       )
       instance_2 = Freeclimb::QueueResult.new(
         uri: "TS",
@@ -277,7 +277,7 @@ describe Freeclimb::QueueResult do
 
         average_queue_removal_time: 1,
 
-        subresource_uris: obj_Object
+        subresource_uris: obj_subresource_uris
       )
       expect(instance_1.eql?(instance_2)).to eq(true)
     end