+++ CHANGELOG.md
@@ -7,7 +7,15 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
-None
+<a name="5.4.0"></a>
+
+## [5.4.0] - 2025-10-24
+
+### Added
+
+- Add models for new TTS models and engine setup
+- Add models and default api methods for Blob API
+
 <a name="5.3.0"></a>
 
 ## [5.3.0] - 2025-09-23